[ rename ] 공통 응답 DTO 경로 변경 ( #1 )
